pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gemini-endpoint-advisor"
version = "0.1.0"
description = "AI-powered endpoint compliance and remediation assistant for Jamf Pro using Google's Gemini."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
  "google-genai",
  "requests",
  "pyyaml",
]

[project.urls]
Homepage = "https://github.com/your-username/gemini-endpoint-advisor"

[project.scripts]
gemini-endpoint-advisor = "gemini_endpoint_advisor.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
